Rename test/utils to test/common
