+++ pom.xml
@@ -307,7 +307,7 @@
     <modules>
         <module>impl</module>
         <module>wildfly-jar</module>
-        <module>tests/utils</module>
+        <module>tests/common</module>
         <module>tests/impl</module>
         <module>tests/wildfly-jar</module>
     </modules>
