+++ tests/common/pom.xml
@@ -10,7 +10,7 @@
         <version>0.2.3.Beta2-SNAPSHOT</version>
     </parent>
 
-    <artifactId>a2a-java-sdk-server-jakarta-test-utils</artifactId>
+    <artifactId>a2a-java-sdk-server-jakarta-test-common</artifactId>
 
     <packaging>jar</packaging>
