+++ tests/common/src/main/resources/arquillian.xml
