+++ tests/impl/src/test/java/org/wildfly/extras/a2a/server/apps/jakarta/JakartaA2AServerTest.java
@@ -16,8 +16,8 @@ import org.jboss.arquillian.container.test.api.RunAsClient;
 import org.jboss.arquillian.junit5.container.annotation.ArquillianTest;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
-import org.wildfly.extras.a2a.server.jakarta.test.utils.A2ATestResource;
-import org.wildfly.extras.a2a.server.jakarta.test.utils.RestApplication;
+import org.wildfly.extras.a2a.server.jakarta.test.common.A2ATestResource;
+import org.wildfly.extras.a2a.server.jakarta.test.common.RestApplication;
 
 
 @ArquillianTest
