+++ tests/wildfly-jar/pom.xml
@@ -27,7 +27,7 @@
         </dependency>
         <dependency>
             <groupId>${project.groupId}</groupId>
-            <artifactId>a2a-java-sdk-server-jakarta-test-utils</artifactId>
+            <artifactId>a2a-java-sdk-server-jakarta-test-common</artifactId>
             <version>${project.version}</version>
             <scope>test</scope>
         </dependency>
